# src/main/java/com/example/forum/models/requests/AuthenticationRequest.java
package com.example.forum.models.requests;

public class AuthenticationRequest {

    private String username;
    private String password;

    public AuthenticationRequest(String username, String password) {
        this.username = username;
        this.password = password;
    }

    public String getUsername() {
        return username;
    }

    public String getPassword() {
        return password;
    }
}